Refactor how static files are served
actix-files doesn't include the files in the binary. This means the binary can't be shipped on its own, it always has to be shipped with the accompanying static folder. This changes how static files are served to use the include_dir crate to embed the files in the binary. Custom routes are then used to retrieve those files.

The way the routes are handled isn't very flexible atm, requiring a new method for each folder. This can be solved in the future though as it isn't a huge issue.