Fix artist image on Safari
Ahhh yes, here we see a wild Safari, being completely broken and not working like every other browser.

Previously, when the viewport got small on Safari, the image would take up the entire background.

This is a little hacky, causing the image to not scale correctly on some very specific screen sizes, and ideally we'll come up with a better solution, but this gets the job done for now.