Fix path_from_url returning incorrect results for invalid URLs
The function would never panic when an invalid URL was pssed in.
Instead, it would simply return and invalid path.
For example, if you passed in `https://google.com`,
it would return `google.com`.

This rewrites the method to instead correctly return an empty string if
an invalid URL is passed in.

Closes #42.